requirements.txt
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
ffmpeg-python==0.2.0
future==1.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
isodate==0.7.2
jsonpath-ng==1.7.0
lxml==5.3.0
outcome==1.3.0.post0
ply==3.11
pycountry==24.6.1
pycryptodome==3.21.0
PySocks==1.7.1
requests==2.32.3
sniffio==1.3.1
sortedcontainers==2.4.0
streamlink==7.0.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
